# small perf improvements and mempool link