## 📝 Description

These are some small and safe changes to improve the performance of fetching bills before we move to a caching solution.

Local rough measurements:
Bill Detail, before: ~500 ms, after: ~70ms
Bill List endpoints (Light, Full, All, Balance, Search) with ~10 full bills: before: ~900 ms, after: ~250 ms

So the change should be noticeable in the UI.

* Switch external time API to premium, only use needed fields and use UTC (bugfix)
* Use local UTC for fetching requests
* Fetch bills in parallel and move computation intensive (deserialization and decryption) tasks to a blocking thread pool, because otherwise the task queue is blocked and the parallelization doesn't help
* Some small linting fixes
* Add mempool link for address to pay

---

## ✅ Checklist

Please ensure the following tasks are completed before requesting a review:

- [x] My code adheres to the coding guidelines of this project.
- [x] I have run `cargo fmt`.
- [x] I have run `cargo clippy`.
- [x] I have added or updated tests (if applicable).
- [x] All CI/CD steps were successful.
- [x] I have updated the documentation (if applicable).
- [x] I have checked that there are no console errors or warnings.
- [x] I have verified that the application builds without errors.
- [x] I've described the changes made to the API. (modification, addition, deletion).

---

## 🚀 Changes Made

See above

---

## 💡 How to Test

Please provide clear instructions on how reviewers can test your changes:

1. Just do bills requests - they should be noticeably faster, with the same functionality as before

---

## 📋 Review Guidelines

Please focus on the following while reviewing:

- [ ] Does the code follow the repository's contribution guidelines?
- [ ] Are there any potential bugs or performance issues?
- [ ] Are there any typos or grammatical errors in the code or comments?